requirements.txt
streamlit
PyMuPDF
pdfplumber
reportlab
Pillow
numpy
pdf2image
PyPDF2
